Change input from int to float

+++ main.py
@@ -31,7 +31,7 @@ label_04.config(padx=5, pady=5)
 
 # Button
 def button_clicked():
-    miles_input = int(input.get())
+    miles_input = float(input.get())
     km_otput = str(1.609344*miles_input)
     output.config(text=km_otput)
